+++ tests/sitemap-index.test.ts
@@ -1,13 +1,19 @@
 import 'babel-polyfill';
-import sm, { EnumChangefreq, EnumYesNo, EnumAllowDeny } from '../index'
-import os from 'os'
-import fs from 'fs'
+import {
+  buildSitemapIndex,
+  createSitemapIndex,
+  EnumChangefreq,
+  EnumYesNo,
+  EnumAllowDeny
+} from '../index'
+import { tmpdir } from 'os'
+import { existsSync, unlinkSync } from 'fs'
 /* eslint-env jest, jasmine */
 function removeFilesArray  (files) {
   if (files && files.length) {
     files.forEach(function (file) {
-      if (fs.existsSync(file)) {
-        fs.unlinkSync(file)
+      if (existsSync(file)) {
+        unlinkSync(file)
       }
     })
   }
@@ -27,7 +33,7 @@ describe('sitemapIndex', () => {
     '</sitemap>' +
     '</sitemapindex>'
 
-    var result = sm.buildSitemapIndex({
+    var result = buildSitemapIndex({
       urls: ['https://test.com/s1.xml', 'https://test.com/s2.xml'],
       xslUrl: 'https://test.com/style.xsl'
     })
@@ -45,7 +51,7 @@ describe('sitemapIndex', () => {
         '</sitemap>' +
     '</sitemapindex>'
 
-    var result = sm.buildSitemapIndex({
+    var result = buildSitemapIndex({
       urls: ['https://test.com/s1.xml', 'https://test.com/s2.xml'],
       xmlNs: 'xmlns="http://www.example.org/schemas/sitemap/0.9"'
     })
@@ -69,7 +75,7 @@ describe('sitemapIndex', () => {
         '</sitemap>' +
     '</sitemapindex>'
 
-    var result = sm.buildSitemapIndex({
+    var result = buildSitemapIndex({
       urls: [
         {
           url: 'https://test.com/s1.xml',
@@ -111,7 +117,7 @@ describe('sitemapIndex', () => {
         '</sitemap>' +
     '</sitemapindex>'
 
-    var result = sm.buildSitemapIndex({
+    var result = buildSitemapIndex({
       urls: [
         {
           url: 'https://test.com/s1.xml'
@@ -133,7 +139,7 @@ describe('sitemapIndex', () => {
         '</sitemap>' +
     '</sitemapindex>'
 
-    var result = sm.buildSitemapIndex({
+    var result = buildSitemapIndex({
       urls: [
         {
           url: 'https://test.com/s1.xml'
@@ -146,7 +152,7 @@ describe('sitemapIndex', () => {
     expect(result).toBe(expectedResult)
   })
   it('simple sitemap index', async () => {
-    const tmp = os.tmpdir()
+    const tmp = tmpdir()
     const url1 = 'http://ya.ru'
     const url2 = 'http://ya2.ru'
     const expectedFiles = [
@@ -157,7 +163,7 @@ describe('sitemapIndex', () => {
 
     expect(
       function () {
-        sm.createSitemapIndex({
+        createSitemapIndex({
           cacheTime: 600000,
           hostname: 'https://www.sitemap.org',
           sitemapName: 'sm-test',
@@ -172,7 +178,7 @@ describe('sitemapIndex', () => {
     removeFilesArray(expectedFiles)
 
     const [err, result] = await new Promise(resolve => {
-      sm.createSitemapIndex({
+      createSitemapIndex({
         cacheTime: 600000,
         hostname: 'https://www.sitemap.org',
         sitemapName: 'sm-test',
@@ -186,21 +192,21 @@ describe('sitemapIndex', () => {
     expect(err).toBeFalsy()
     expect(result).toBe(true)
     expectedFiles.forEach(function (expectedFile) {
-      expect(fs.existsSync(expectedFile)).toBe(true)
+      expect(existsSync(expectedFile)).toBe(true)
     })
   })
   it('sitemap without callback', () => {
-    sm.createSitemapIndex({
+    createSitemapIndex({
       cacheTime: 600000,
       hostname: 'http://www.sitemap.org',
       sitemapName: 'sm-test',
       sitemapSize: 1,
-      targetFolder: os.tmpdir(),
+      targetFolder: tmpdir(),
       urls: ['http://ya.ru', 'http://ya2.ru']
     })
   })
   it('sitemap with gzip files', async () => {
-    const tmp = os.tmpdir()
+    const tmp = tmpdir()
     const url1 = 'http://ya.ru'
     const url2 = 'http://ya2.ru'
     const expectedFiles = [
@@ -213,7 +219,7 @@ describe('sitemapIndex', () => {
     removeFilesArray(expectedFiles)
 
     const [err, result] = await new Promise(resolve => {
-      sm.createSitemapIndex({
+      createSitemapIndex({
         cacheTime: 600000,
         hostname: 'http://www.sitemap.org',
         sitemapName: 'sm-test',
@@ -227,7 +233,7 @@ describe('sitemapIndex', () => {
     expect(err).toBeFalsy()
     expect(result).toBe(true)
     expectedFiles.forEach(function (expectedFile) {
-      expect(fs.existsSync(expectedFile)).toBe(true)
+      expect(existsSync(expectedFile)).toBe(true)
     })
   })
 })