+++ tests/sitemap.test.ts
@@ -6,8 +6,8 @@
 /* eslint-env jest, jasmine */
 import 'babel-polyfill'
 
-import sm, { EnumChangefreq, EnumYesNo, EnumAllowDeny } from '../index'
-import zlib from 'zlib'
+import { Sitemap, createSitemap, EnumChangefreq, EnumYesNo, EnumAllowDeny } from '../index'
+import { gzipSync, gunzipSync } from 'zlib'
 
 const urlset = '<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" ' +
              'xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" ' +
@@ -22,22 +22,13 @@ const xmlPriority = '<priority>0.9</priority>'
 const xmlLoc = '<loc>http://ya.ru/</loc>'
 
 describe('sitemap', () => {
-  it('sitemap empty urls', () => {
-    const smEmpty = new sm.Sitemap()
-
-    expect(smEmpty.urls).toEqual([])
-  })
-
-  it('sitemap.urls is an array', () => {
-    const url = 'ya.ru'
-    const smOne = new sm.Sitemap(url)
-
-    expect(smOne.urls).toEqual([url])
+  it('can be instantiated without options', () => {
+    expect(() => (new Sitemap())).not.toThrow()
   })
 
   it('simple sitemap', () => {
     var url = 'http://ya.ru'
-    var ssp = new sm.Sitemap()
+    var ssp = new Sitemap()
     ssp.add(url)
 
     expect(ssp.toString()).toBe(
@@ -52,7 +43,7 @@ describe('sitemap', () => {
     it('accepts url strings', () => {
       var url = '/some_page'
       let hostname = 'http://ya.ru'
-      var ssp = new sm.Sitemap(undefined, hostname)
+      var ssp = new Sitemap({hostname})
       ssp.add(url)
 
       expect(ssp.toString()).toBe(
@@ -65,7 +56,7 @@ describe('sitemap', () => {
     })
     it('accepts config url objects', () => {
       var url = 'http://ya.ru'
-      var ssp = new sm.Sitemap()
+      var ssp = new Sitemap()
       ssp.add({ url, changefreq: EnumChangefreq.DAILY })
 
       expect(ssp.toString()).toBe(
@@ -81,7 +72,7 @@ describe('sitemap', () => {
 
   it('encodes URLs', () => {
     var url = 'http://ya.ru/?foo=bar baz'
-    var ssp = new sm.Sitemap()
+    var ssp = new Sitemap()
     ssp.add(url)
 
     expect(ssp.toString()).toBe(
@@ -95,7 +86,7 @@ describe('sitemap', () => {
 
   it('simple sitemap with dynamic xmlNs', () => {
     var url = 'http://ya.ru'
-    var ssp = sm.createSitemap({
+    var ssp = createSitemap({
       xmlNs: 'xmlns="http://www.sitemaps.org/schemas/sitemap/0.9"'
     })
     ssp.add(url)
@@ -108,27 +99,9 @@ describe('sitemap', () => {
     '</urlset>')
   })
 
-  it('simple sitemap toXML async with two callback arguments', async () => {
-    var url = 'http://ya.ru'
-    var ssp = new sm.Sitemap()
-    ssp.add(url)
-
-    const [ err, xml ] = await new Promise(resolve => {
-      ssp.toXML((...args) => { resolve(args) })
-    })
-    expect(err).toBeUndefined()
-    expect(xml).toBe(
-      xmlDef +
-              urlset +
-                '<url>' +
-                    xmlLoc +
-                '</url>' +
-              '</urlset>')
-  })
-
   it('simple sitemap toXML sync', () => {
     var url = 'http://ya.ru'
-    var ssp = new sm.Sitemap()
+    var ssp = new Sitemap()
     ssp.add(url)
 
     expect(ssp.toXML()).toBe(
@@ -141,10 +114,10 @@ describe('sitemap', () => {
   })
 
   it('simple sitemap toGzip sync', () => {
-    var ssp = new sm.Sitemap()
+    var ssp = new Sitemap()
     ssp.add('http://ya.ru')
 
-    expect(ssp.toGzip()).toEqual(zlib.gzipSync(
+    expect(ssp.toGzip()).toEqual(gzipSync(
       xmlDef +
               urlset +
                 '<url>' +
@@ -155,12 +128,12 @@ describe('sitemap', () => {
   })
 
   it('simple sitemap toGzip async', () => {
-    var ssp = new sm.Sitemap()
+    var ssp = new Sitemap()
     ssp.add('http://ya.ru')
 
     ssp.toGzip(function (error, result) {
       expect(error).toBe(null)
-      expect(zlib.gunzipSync(result).toString()).toBe(
+      expect(gunzipSync(result).toString()).toBe(
         xmlDef +
             urlset +
             '<url>' +
@@ -172,7 +145,7 @@ describe('sitemap', () => {
   })
 
   it('video attributes', () => {
-    var smap = sm.createSitemap({
+    var smap = createSitemap({
       urls: [
         {
           'url': 'https://roosterteeth.com/episode/achievement-hunter-achievement-hunter-burnout-paradise-millionaires-club',
@@ -224,7 +197,7 @@ describe('sitemap', () => {
   })
 
   it('sitemap: hostname, createSitemap', () => {
-    var smap = sm.createSitemap({
+    var smap = createSitemap({
       hostname: 'http://test.com',
       urls: [
         { url: '/', changefreq: EnumChangefreq.ALWAYS, priority: 1 },
@@ -270,7 +243,7 @@ describe('sitemap', () => {
   })
 
   it('custom xslUrl', () => {
-    var smap = sm.createSitemap({
+    var smap = createSitemap({
       urls: [
         { url: 'http://test.com/', changefreq: EnumChangefreq.ALWAYS, priority: 1 }
       ],
@@ -292,8 +265,9 @@ describe('sitemap', () => {
   it('sitemap: invalid changefreq error', () => {
     expect(
       function () {
-        sm.createSitemap({
+        createSitemap({
           hostname: 'http://test.com',
+          // @ts-ignore
           urls: [{ url: '/', changefreq: 'allllways' }]
         }).toString()
       }
@@ -302,7 +276,7 @@ describe('sitemap', () => {
   it('sitemap: invalid priority error', () => {
     expect(
       function () {
-        sm.createSitemap({
+        createSitemap({
           hostname: 'http://test.com',
           urls: [{ url: '/', priority: 1.1 }]
         }).toString()
@@ -310,7 +284,7 @@ describe('sitemap', () => {
     ).toThrowError(/priority is invalid/)
   })
   it('sitemap: test cache', () => {
-    const smap = sm.createSitemap({
+    const smap = createSitemap({
       hostname: 'http://test.com',
       cacheTime: 500, // 0.5 sec
       urls: [
@@ -352,7 +326,7 @@ describe('sitemap', () => {
     }, 1000)
   })
   it('sitemap: test cache off', () => {
-    const smap = sm.createSitemap({
+    const smap = createSitemap({
       hostname: 'http://test.com',
       // cacheTime: 0,  // cache disabled
       urls: [
@@ -386,7 +360,7 @@ describe('sitemap', () => {
               '</urlset>')
   })
   it('sitemap: handle urls with "http" in the path', () => {
-    var smap = sm.createSitemap({
+    var smap = createSitemap({
       hostname: 'http://test.com',
       urls: [
         { url: '/page-that-mentions-http:-in-the-url/', changefreq: EnumChangefreq.WEEKLY, priority: 0.3 }
@@ -404,7 +378,7 @@ describe('sitemap', () => {
     expect(smap.toString()).toBe(xml)
   })
   it('sitemap: handle urls with "&" in the path', () => {
-    var smap = sm.createSitemap({
+    var smap = createSitemap({
       hostname: 'http://test.com',
       urls: [
         { url: '/page-that-mentions-&-in-the-url/', changefreq: EnumChangefreq.WEEKLY, priority: 0.3 }
@@ -422,7 +396,7 @@ describe('sitemap', () => {
     expect(smap.toString()).toBe(xml)
   })
   it('sitemap: keep urls that start with http:// or https://', () => {
-    const smap = sm.createSitemap({
+    const smap = createSitemap({
       hostname: 'http://test.com',
       urls: [
         { url: 'http://ya.ru/page-1/', changefreq: EnumChangefreq.WEEKLY, priority: 0.3 },
@@ -446,10 +420,10 @@ describe('sitemap', () => {
     expect(smap.toString()).toBe(xml)
   })
   it('sitemap: del by string', () => {
-    const smap = sm.createSitemap({
+    const smap = createSitemap({
       hostname: 'http://test.com',
       urls: [
-        { url: 'http://ya.ru/page-1/', changefreq: EnumChangefreq.WEEKLY, priority: 0.3 },
+        { url: '/page-1/', changefreq: EnumChangefreq.WEEKLY, priority: 0.3 },
         { url: 'https://ya.ru/page-2/', changefreq: EnumChangefreq.WEEKLY, priority: 0.3 }
       ]
     })
@@ -461,12 +435,12 @@ describe('sitemap', () => {
                     '<priority>0.3</priority>' +
                 '</url>' +
               '</urlset>'
-    smap.del('http://ya.ru/page-1/')
+    smap.del('/page-1/')
 
     expect(smap.toString()).toBe(xml)
   })
   it('sitemap: del by object', () => {
-    const smap = sm.createSitemap({
+    const smap = createSitemap({
       hostname: 'http://test.com',
       urls: [
         { url: 'http://ya.ru/page-1/', changefreq: EnumChangefreq.WEEKLY, priority: 0.3 },
@@ -487,16 +461,22 @@ describe('sitemap', () => {
   })
   it('test for #27', () => {
     var staticUrls = ['/', '/terms', '/login']
-    var sitemap = sm.createSitemap({ urls: staticUrls })
+    var sitemap = createSitemap({ urls: staticUrls, hostname: 'http://example.com' })
     sitemap.add({ url: '/details/' + 'url1' })
 
-    var sitemap2 = sm.createSitemap({ urls: staticUrls })
+    var sitemap2 = createSitemap({ urls: staticUrls, hostname: 'http://example.com'})
 
-    expect(sitemap.urls).toEqual(['/', '/terms', '/login', { url: '/details/url1' }])
-    expect(sitemap2.urls).toEqual(['/', '/terms', '/login'])
+    expect(sitemap.contains({url: 'http://example.com/'})).toBeTruthy()
+    expect(sitemap.contains({url: 'http://example.com/terms'})).toBeTruthy()
+    expect(sitemap.contains({url: 'http://example.com/login'})).toBeTruthy()
+    expect(sitemap.contains({url: 'http://example.com/details/url1'})).toBeTruthy()
+    expect(sitemap2.contains({url: 'http://example.com/'})).toBeTruthy()
+    expect(sitemap2.contains({url: 'http://example.com/terms'})).toBeTruthy()
+    expect(sitemap2.contains({url: 'http://example.com/login'})).toBeTruthy()
+    expect(sitemap2.contains({url: 'http://example.com/details/url1'})).toBeFalsy()
   })
   it('sitemap: langs', () => {
-    var smap = sm.createSitemap({
+    var smap = createSitemap({
       urls: [
         { url: 'http://test.com/page-1/',
           changefreq: EnumChangefreq.WEEKLY,
@@ -520,22 +500,13 @@ describe('sitemap', () => {
               '</urlset>')
   })
   it('sitemap: normalize urls, see #39', async () => {
-    const [xml1, xml2] = await Promise.all(
-      ['http://ya.ru', 'http://ya.ru/'].map(function (hostname) {
-        var ssp = new sm.Sitemap(null, hostname)
-        ssp.add('page1')
-        ssp.add('/page2')
+    const [xml1, xml2] = ['http://ya.ru', 'http://ya.ru/'].map(function (hostname) {
+      var ssp = new Sitemap({hostname})
+      ssp.add('page1')
+      ssp.add('/page2')
 
-        return new Promise(resolve => {
-          ssp.toXML(function (err, xml) {
-            if (err) {
-              console.error(err)
-            }
-            resolve(xml)
-          })
-        })
-      })
-    )
+      return ssp.toXML()
+    })
     expect(xml1).toBe(xml2)
     expect(xml1).toBe(
       xmlDef +
@@ -549,7 +520,7 @@ describe('sitemap', () => {
       '</urlset>')
   })
   it('sitemap: langs with hostname', () => {
-    var smap = sm.createSitemap({
+    var smap = createSitemap({
       hostname: 'http://test.com',
       urls: [
         { url: '/page-1/',
@@ -573,21 +544,8 @@ describe('sitemap', () => {
                 '</url>' +
               '</urlset>')
   })
-  it('sitemap: error thrown in async-style .toXML()', () => {
-    var smap = sm.createSitemap({
-      hostname: 'http://test.com',
-      urls: [
-        { url: '/page-1/', changefreq: EnumChangefreq.WEEKLY, priority: 0.3 }
-      ]
-    })
-    var error = new Error('Some error happens')
-    smap.toString = () => { throw error }
-    smap.toXML(function (err, xml) {
-      expect(err).toBe(error)
-    })
-  })
   it('sitemap: android app linking', () => {
-    var smap = sm.createSitemap({
+    var smap = createSitemap({
       urls: [
         { url: 'http://test.com/page-1/',
           changefreq: EnumChangefreq.WEEKLY,
@@ -607,7 +565,7 @@ describe('sitemap', () => {
               '</urlset>')
   })
   it('sitemap: AMP', () => {
-    var smap = sm.createSitemap({
+    var smap = createSitemap({
       urls: [
         { url: 'http://test.com/page-1/',
           changefreq: EnumChangefreq.WEEKLY,
@@ -626,12 +584,12 @@ describe('sitemap', () => {
       '</urlset>')
   })
   it('sitemap: expires', () => {
-    var smap = sm.createSitemap({
+    var smap = createSitemap({
       urls: [
         { url: 'http://test.com/page-1/',
           changefreq: EnumChangefreq.WEEKLY,
           priority: 0.3,
-          expires: new Date('2016-09-13') }
+          expires: new Date('2016-09-13').toString() }
       ]
     })
     expect(smap.toString()).toBe(
@@ -645,7 +603,7 @@ describe('sitemap', () => {
       '</urlset>')
   })
   it('sitemap: image with caption', () => {
-    var smap = sm.createSitemap({
+    var smap = createSitemap({
       hostname: 'http://test.com',
       urls: [
         { url: '/a', img: { url: '/image.jpg?param&otherparam', caption: 'Test Caption' } }
@@ -665,7 +623,7 @@ describe('sitemap', () => {
       '</urlset>')
   })
   it('sitemap: image with caption, title, geo_location, license', () => {
-    var smap = sm.createSitemap({
+    var smap = createSitemap({
       urls: [
         { url: 'http://test.com',
           img: {
@@ -695,7 +653,7 @@ describe('sitemap', () => {
       '</urlset>')
   })
   it('sitemap: images with captions', () => {
-    var smap = sm.createSitemap({
+    var smap = createSitemap({
       urls: [
         { url: 'http://test.com', img: { url: 'http://test.com/image.jpg', caption: 'Test Caption' } },
         { url: 'http://test.com/page2/', img: { url: 'http://test.com/image2.jpg', caption: 'Test Caption 2' } }
@@ -722,7 +680,7 @@ describe('sitemap', () => {
       '</urlset>')
   })
   it('sitemap: images with captions add', () => {
-    var smap = sm.createSitemap({
+    var smap = createSitemap({
       hostname: 'http://test.com',
       urls: [
         {
@@ -735,7 +693,7 @@ describe('sitemap', () => {
       ]
     })
 
-    smap.urls.push({ url: '/index2.html', img: [{ url: '/image3.jpg', caption: 'Test Caption 3' }] })
+    smap.add({ url: '/index2.html', img: [{ url: '/image3.jpg', caption: 'Test Caption 3' }] })
 
     expect(smap.toString()).toBe(
       xmlDef +
@@ -761,7 +719,7 @@ describe('sitemap', () => {
       '</urlset>')
   })
   it('sitemap: video', () => {
-    var smap = sm.createSitemap({
+    var smap = createSitemap({
       urls: [
         {
           'url': 'https://roosterteeth.com/episode/achievement-hunter-achievement-hunter-burnout-paradise-millionaires-club',